Move CheckpointStatement to tcl package and visitor logic to tcl visitor, remove LockTableStatement
