+++ parser/sql/dialect/opengauss/src/main/antlr4/imports/opengauss/TCLStatement.g4
@@ -84,3 +84,21 @@ constraintsSetList
     : ALL | qualifiedNameList
     ;
 
+checkpoint
+    : CHECKPOINT
+    ;
+
+lock
+    : LOCK TABLE? relationExprList (IN lockType MODE)? NOWAIT?
+    ;
+
+lockType
+    : ACCESS SHARE
+    | ROW SHARE
+    | ROW EXCLUSIVE
+    | SHARE UPDATE EXCLUSIVE
+    | SHARE
+    | SHARE ROW EXCLUSIVE
+    | EXCLUSIVE
+    | ACCESS EXCLUSIVE
+    ;
