+++ parser/sql/dialect/postgresql/src/main/antlr4/imports/postgresql/TCLStatement.g4
@@ -102,3 +102,7 @@ lockType
 prepareTransaction
     : PREPARE TRANSACTION STRING_
     ;
+
+checkpoint
+    : CHECKPOINT
+    ;
